+++ src/account_readers/read_pools_trace_account.rs
@@ -10,7 +10,7 @@ impl DvlReadable for PoolsTraceAccount {
 
     fn read<'a>(reader: &DvlAccountReader, params: Self::AdditionalCheckParams<'a>) -> Result<Box<Self>, Box<dyn Error>> where Self: Sized {
         let workers_account = reader.read::<AllWorkersAccount>(()).unwrap();
-        let worker = workers_account.workers[params.id];
+        let worker = workers_account.workers[params.id as usize];
         let public_key = &worker.pools_trace_address;
         // let account =  Self::read_by_public_key(reader, public_key)?;
         let mut rpc_data = reader.client.get_account(public_key)?;
